用户服务-event_outbox表、添加或取消收藏、插入事件

+++ heima-leadnews/heima-leadnews-service/heima-leadnews-user/src/main/java/com/heima/user/interceptor/AppTokenInterceptor.java
@@ -26,7 +26,7 @@ public class AppTokenInterceptor implements HandlerInterceptor {
         String userId = request.getHeader("userId");
         if (userId != null){
             ApUser apUser = new ApUser();
-            apUser.setId(Integer.parseInt(userId));
+            apUser.setId(Long.parseLong(userId));
             AppThreadLocalUtil.setUser(apUser);
         }
